Разбивка main.js на модули
