+++ index.html
@@ -10,7 +10,7 @@
 
   <!-- Scripts -->
   <script async src="/js/functions.js"></script>
-  <script async src="/js/main.js"></script>
+  <script async src="/js/main.js" type="module"></script>
 </head>
 
 <body>
